Resolve nodes resulting from a GetRelatedDefinitions() call.
Earlier, GetRelatedDefinitions() would return an abstract XrefNode that
didn't have any cross reference metadata. This change causes an
additional HAS_DEFINITION cross-reference lookup to be performed so that
the resulting XrefNode corresponds to the definition of the symbol.

In addition, fixes handling of long URLs by splitting the query
parameters into the body.