Breaking API changes in responses to CodeSearch backend migration.
`https://cs.chromium.org` went through a substantial change at the end
of 2017 which involved multiple behavior changes to the JSON RPC
endpoints used by `codesearch-py`. These resulted in breaking API
changes as described below.

There's no real supported API for external libraries such as this one to
communicate with the `https://cs.chromium.org` backend. Instead, this
library relies on a set of reverse-engineered JSON endpoints that are
meant to support the CodeSearch web UI. Chrome Infra is aware of this
use, but is in no way required to support it. Hence breakages such as
what was observed during the Grok->Kythe backened migration are expected
and supposed to be dealt with the `codesearch-py` and `vim-codesearch`
maintainers.

The changes that were required are summarized in
`doc/changes_since_grok_migration.md`.

Addresses https://github.com/chromium/vim-codesearch/issues/8 once
`vim-codesearch` is updated accordingly.