Support CALLED_BY references from XrefNode.
One of the features that we lost during the Kythe migration is the
ability to lookup call sites for a signature directly from the
xref_search request. Instaed now we need to do a call graph request. To
make it easier for consumers, this change re-adds the CALLED_BY request
type to XrefNode.Traverse().

In addition, this change adds first class support for CodeBlocks. This
would compensate for the missing DECLARES et. al. functionality that was
also removed during the Kythe migration.